+++ packages/ledger-icrc/src/nft-ledger.canister.spec.ts
@@ -1,0 +1,41 @@
+import type { ActorSubclass } from "@icp-sdk/core/agent";
+import { mock } from "vitest-mock-extended";
+import type { _SERVICE as IcrcNftLedgerService } from "./candid/icrc_nft-ledger";
+import {
+  ledgerCanisterIdMock,
+  tokenMetadataResponseMock,
+} from "./mocks/ledger.mock";
+import { IcrcNftLedgerCanister } from "./nft-ledger.canister";
+
+describe("NFT Ledger canister", () => {
+  describe("collectionMetadata", () => {
+    it("should return the collection metadata", async () => {
+      const service = mock<ActorSubclass<IcrcNftLedgerService>>();
+      service.icrc7_collection_metadata.mockResolvedValue(
+        tokenMetadataResponseMock,
+      );
+
+      const canister = IcrcNftLedgerCanister.create({
+        canisterId: ledgerCanisterIdMock,
+        certifiedServiceOverride: service,
+      });
+
+      const res = await canister.collectionMetadata({});
+
+      expect(res).toEqual(tokenMetadataResponseMock);
+    });
+
+    it("should fail when the canister call fails", async () => {
+      const mockError = new Error("Canister call failed");
+      const service = mock<ActorSubclass<IcrcNftLedgerService>>();
+      service.icrc7_collection_metadata.mockRejectedValue(mockError);
+
+      const canister = IcrcNftLedgerCanister.create({
+        canisterId: ledgerCanisterIdMock,
+        certifiedServiceOverride: service,
+      });
+
+      await expect(canister.collectionMetadata({})).rejects.toThrow(mockError);
+    });
+  });
+});